pyproject.toml
[project]
name = "referral-platform"
version = "0.1.0"
description = ""
authors = [
    {name = "Alexey",email = "rozanov7475@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-phonenumber-field (>=8.1.0,<9.0.0)",
    "phonenumbers (>=9.0.10,<10.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
]

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.36.0"
black = "^25.1.0"
flake8 = "^7.3.0"
pytest = "^8.4.1"
faker = "^37.4.2"
pytest-django = "^4.11.1"

[tool.poe.env]
PYTHONPATH = "src"
APP_NAME = "cafe-order-system"
DC = "docker compose -p ${APP_NAME} -f docker-compose.dev.yml"
DC_PROD = "docker compose -p ${APP_NAME} -f docker-compose.prod.yml"

[tool.poe.tasks]
up = "$DC up"
down = "$DC down"
test = "pytest"
start = "gunicorn config.wsgi:application --bind 0.0.0.0:8084"
collectstatic = { shell = "cd src && python3 manage.py collectstatic --noinput" }
runserver = { shell = "cd src && python manage.py runserver" }
migrate = { shell = "cd src && python manage.py migrate" }
makemigrations = { shell = "cd src && python manage.py makemigrations" }
up_prod = "$DC_PROD up"
down_prod = "$DC_PROD down"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = "tests.py test_*.py *_tests.py"
addopts = ["-rsxX", "-l", "--tb=short", "--strict-markers"]
markers = ["unit", "integration", "e2e"]
cache_dir = ".venv/cache/pytest"

[tool.black]
line-length = 79

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
